[PM-5693] Migrate SDK to KeyStore (#8)
## 🎟️ Tracking

https://bitwarden.atlassian.net/browse/PM-5693

## 📔 Objective

Migrated from https://github.com/bitwarden/sdk/pull/1117

Migrate the codebase to the new KeyStore introduced in
https://github.com/bitwarden/sdk-internal/pull/7.

EncryptionSettings was removed from Client, though it is still used to
initialize the KeyStore. Ideally we'd move that initialization code over
directly into the crypto crate but this PR is big enough as it is.

There are still some things using keys directly and
KeyEncryptable/KeyDecryptable, like MasterKey, PinKey, DeviceKey,
private key fingerprint. Those would need to be migrated over on a
separate PR.

We need to remove the rest of the uses of SymmetricCryptoKey from the
client crates, then we can remove the internal boxing they are doing, as
the keys would be protected by the KeyStore instead.


Secrets Manager code should be moved to using the
Encryptable/Decryptable interface, as it's encrypting fields one by one,
I'll look into doing that on a separate PR.

## ⏰ Reminders before review

- Contributor guidelines followed
- All formatters and local linters executed and passed
- Written new unit and / or integration tests where applicable
- Protected functional changes with optionality (feature flags)
- Used internationalization (i18n) for all UI strings
- CI builds passed
- Communicated to DevOps any deployment requirements
- Updated any necessary documentation (Confluence, contributing docs) or
informed the documentation
  team

## 🦮 Reviewer guidelines

<!-- Suggested interactions but feel free to use (or not) as you desire!
-->

- 👍 (`:+1:`) or similar for great changes
- 📝 (`:memo:`) or ℹ️ (`:information_source:`) for notes or general info
- ❓ (`:question:`) for questions
- 🤔 (`:thinking:`) or 💭 (`:thought_balloon:`) for more open inquiry
that's not quite a confirmed
  issue and could potentially benefit from discussion
- 🎨 (`:art:`) for suggestions / improvements
- ❌ (`:x:`) or ⚠️ (`:warning:`) for more significant problems or
concerns needing attention
- 🌱 (`:seedling:`) or ♻️ (`:recycle:`) for future improvements or
indications of technical debt
- ⛏ (`:pick:`) for minor or nitpick changes