Build uniffi demo apps in CI (#162)
## 🎟️ Tracking

<!-- Paste the link to the Jira or GitHub issue or otherwise describe /
point to where this change is coming from. -->

## 📔 Objective

We've had issues (specially on iOS) where some reasonable looking
changes breaks the use of the library, for some reason those build
errors would only appear when trying to compile an app with the SDK and
not when just building the SDK. This is happening again now with the
duplicate SSHKey item, which is fixed in #160, so this PR is based on
top of that one.

This PR does a couple of things to improve this:
- Enabled the building of the swift SDK in pull requests (note that the
release is already behind a `if: github.ref == 'refs/heads/main'`)
- Added a step building the swift demo app, which should help in
surfacing those errors in the future
- Also added a step building the kotlin demo app, we haven't seen any
problems with it in the past, but it would help ensure the demo app is
up to date at least.

## ⏰ Reminders before review

- Contributor guidelines followed
- All formatters and local linters executed and passed
- Written new unit and / or integration tests where applicable
- Protected functional changes with optionality (feature flags)
- Used internationalization (i18n) for all UI strings
- CI builds passed
- Communicated to DevOps any deployment requirements
- Updated any necessary documentation (Confluence, contributing docs) or
informed the documentation
  team

## 🦮 Reviewer guidelines

<!-- Suggested interactions but feel free to use (or not) as you desire!
-->

- 👍 (`:+1:`) or similar for great changes
- 📝 (`:memo:`) or ℹ️ (`:information_source:`) for notes or general info
- ❓ (`:question:`) for questions
- 🤔 (`:thinking:`) or 💭 (`:thought_balloon:`) for more open inquiry
that's not quite a confirmed
  issue and could potentially benefit from discussion
- 🎨 (`:art:`) for suggestions / improvements
- ❌ (`:x:`) or ⚠️ (`:warning:`) for more significant problems or
concerns needing attention
- 🌱 (`:seedling:`) or ♻️ (`:recycle:`) for future improvements or
indications of technical debt
- ⛏ (`:pick:`) for minor or nitpick changes

---------

Co-authored-by: Bernd Schoolmann <mail@quexten.com>