[PM-17370] Refactor symmetric keys to use enum (#127)
## 🎟️ Tracking

https://bitwarden.atlassian.net/browse/PM-17370

## 📔 Objective

Symmetric crypto keys (in general) are just random arrays of data - in
contrast to asymmetric keys which require specific structure. Still, it
is cryptographic best practice to split these keys apart entirely, and
not have them confused. An algorithm switch should include a change to a
new key type.

For the upcoming xchacha cipher, we will introduce a symmetric key that
is the same size as the old hmac-less cbc type 0 cipher. In the sdk we
currently just represent this as one key with two byte arrays, "key" and
"mac key" which is confusing and could lead to mis-usage, which can lead
to bugs or even cryptographic flaws.

For this reason, it is paramount to enforce the key types in the type
system. This PR refactors the current key usage to represent symmetric
keys as an enum, with the two types being the current encstring type 0
and type 2, each of which has it's own struct with the data it needs.

In a later PR, the serialization format for symmetric keys will be
adjusted to include the cipher type, so keys are not categorized based
on the size alone.

## ⏰ Reminders before review

- Contributor guidelines followed
- All formatters and local linters executed and passed
- Written new unit and / or integration tests where applicable
- Protected functional changes with optionality (feature flags)
- Used internationalization (i18n) for all UI strings
- CI builds passed
- Communicated to DevOps any deployment requirements
- Updated any necessary documentation (Confluence, contributing docs) or
informed the documentation
  team

## 🦮 Reviewer guidelines

<!-- Suggested interactions but feel free to use (or not) as you desire!
-->

- 👍 (`:+1:`) or similar for great changes
- 📝 (`:memo:`) or ℹ️ (`:information_source:`) for notes or general info
- ❓ (`:question:`) for questions
- 🤔 (`:thinking:`) or 💭 (`:thought_balloon:`) for more open inquiry
that's not quite a confirmed
  issue and could potentially benefit from discussion
- 🎨 (`:art:`) for suggestions / improvements
- ❌ (`:x:`) or ⚠️ (`:warning:`) for more significant problems or
concerns needing attention
- 🌱 (`:seedling:`) or ♻️ (`:recycle:`) for future improvements or
indications of technical debt
- ⛏ (`:pick:`) for minor or nitpick changes

---------

Co-authored-by: Daniel García <dani-garcia@users.noreply.github.com>
Co-authored-by: Oscar Hinton <Hinton@users.noreply.github.com>