Always use WASM MVP (#182)
## 🎟️ Tracking

## 📔 Objective

Webpack seems to be having some compatibility issues with the reference
types feature of WASM that was enabled with Rust 1.82.0, leading to some
strange results:
The `verify_asymmetric_keys` function should be returning two booleans:
```
{privateKeyDecryptable: true, validPrivateKey: true}
```
and is instead returning two strings with value `identityUrl`:
```
{privateKeyDecryptable: "identityUrl", validPrivateKey: "identityUrl"}
```
I've tried to reproduce the error in a Jest test and I can't, which is
what leads me to blame webpack. By using the MVP builds for both the
WASM artifact and the WASM2JS build it seems to fix the issue for me.

## ⏰ Reminders before review

- Contributor guidelines followed
- All formatters and local linters executed and passed
- Written new unit and / or integration tests where applicable
- Protected functional changes with optionality (feature flags)
- Used internationalization (i18n) for all UI strings
- CI builds passed
- Communicated to DevOps any deployment requirements
- Updated any necessary documentation (Confluence, contributing docs) or
informed the documentation
  team

## 🦮 Reviewer guidelines

<!-- Suggested interactions but feel free to use (or not) as you desire!
-->

- 👍 (`:+1:`) or similar for great changes
- 📝 (`:memo:`) or ℹ️ (`:information_source:`) for notes or general info
- ❓ (`:question:`) for questions
- 🤔 (`:thinking:`) or 💭 (`:thought_balloon:`) for more open inquiry
that's not quite a confirmed
  issue and could potentially benefit from discussion
- 🎨 (`:art:`) for suggestions / improvements
- ❌ (`:x:`) or ⚠️ (`:warning:`) for more significant problems or
concerns needing attention
- 🌱 (`:seedling:`) or ♻️ (`:recycle:`) for future improvements or
indications of technical debt
- ⛏ (`:pick:`) for minor or nitpick changes