Add readmes to all crates (#200)
In an effort to improve documentation for our public crates this PR adds
relevant boilerplate readmes to every crate and moves the disclaimer to
the workspace.

Uses `readme = "DISCLAIMER.md"` to configure which readme is shown on
crates.io. This allows us to use `README.md` for internal documentation.
The READMEs are also included in rustdoc in an effort to avoid
duplicating documentation.