[PM-18772] Rename AesCbc256 to Aes256Cbc (#172)
## 🎟️ Tracking

https://bitwarden.atlassian.net/browse/PM-18772
Clients https://github.com/bitwarden/clients/pull/13637

## 📔 Objective

Currently we use: `Aes256Cbc`, but also `AesCbc256` in the SDK. In the
web clients we use `AesCbc256`. In general the former order is more
common.

For instance, TLS defines: `TLS_DHE_RSA_WITH_AES_256_CBC_SHA384`. JWK
defines: `AES_128_CBC_HMAC_SHA_256`, SSH defines `aes256-cbc`, COSE
defines: `A256CBC`.
This is because the block cipher is: AES-128-Bit / AES-256-Bit, and the
mode of operation is `CBC`. Semantically, we are currently splitting up
the block cipher used and interjecting the mode of operation in the
middle.

This PR renames the instances used in the SDK of `AesCbc256` to be
consistent with other standards, and to be consistent within our
codebase. There will be a corresponding PR for the web clients (this is
in draft until then).

## ⏰ Reminders before review

- Contributor guidelines followed
- All formatters and local linters executed and passed
- Written new unit and / or integration tests where applicable
- Protected functional changes with optionality (feature flags)
- Used internationalization (i18n) for all UI strings
- CI builds passed
- Communicated to DevOps any deployment requirements
- Updated any necessary documentation (Confluence, contributing docs) or
informed the documentation
  team

## 🦮 Reviewer guidelines

<!-- Suggested interactions but feel free to use (or not) as you desire!
-->

- 👍 (`:+1:`) or similar for great changes
- 📝 (`:memo:`) or ℹ️ (`:information_source:`) for notes or general info
- ❓ (`:question:`) for questions
- 🤔 (`:thinking:`) or 💭 (`:thought_balloon:`) for more open inquiry
that's not quite a confirmed
  issue and could potentially benefit from discussion
- 🎨 (`:art:`) for suggestions / improvements
- ❌ (`:x:`) or ⚠️ (`:warning:`) for more significant problems or
concerns needing attention
- 🌱 (`:seedling:`) or ♻️ (`:recycle:`) for future improvements or
indications of technical debt
- ⛏ (`:pick:`) for minor or nitpick changes