[PM-19966] Remove subclient lifetimes (#198)
## 🎟️ Tracking

https://bitwarden.atlassian.net/browse/PM-19966

## 📔 Objective

The subclient lifetimes have been a pain point that cause us to need to
have create our duplicate copies of the structs for WASM and Uniffi that
don't deal with them.

If instead of doing that we wrap the `Client`s `InternalClient` in
`Arc`, we can just cheaply clone the `Arc` and remove all the lifetimes
from the subclients. This opens up the posibility of using the same API
for Rust/WASM/UniFFI in the future, and also simplifies the bindings a
bit in the short term.

This does have the negative that each subclient access requires an
atomic operation clone, but that's negligible for our usecases. (And we
were already doing basically the same thing for Uniffi and WASM anyway)

## ⏰ Reminders before review

- Contributor guidelines followed
- All formatters and local linters executed and passed
- Written new unit and / or integration tests where applicable
- Protected functional changes with optionality (feature flags)
- Used internationalization (i18n) for all UI strings
- CI builds passed
- Communicated to DevOps any deployment requirements
- Updated any necessary documentation (Confluence, contributing docs) or
informed the documentation
  team

## 🦮 Reviewer guidelines

<!-- Suggested interactions but feel free to use (or not) as you desire!
-->

- 👍 (`:+1:`) or similar for great changes
- 📝 (`:memo:`) or ℹ️ (`:information_source:`) for notes or general info
- ❓ (`:question:`) for questions
- 🤔 (`:thinking:`) or 💭 (`:thought_balloon:`) for more open inquiry
that's not quite a confirmed
  issue and could potentially benefit from discussion
- 🎨 (`:art:`) for suggestions / improvements
- ❌ (`:x:`) or ⚠️ (`:warning:`) for more significant problems or
concerns needing attention
- 🌱 (`:seedling:`) or ♻️ (`:recycle:`) for future improvements or
indications of technical debt
- ⛏ (`:pick:`) for minor or nitpick changes