[PM-19603] Asymmetric encrypt/encapsulate interface (#202)
## 🎟️ Tracking

https://bitwarden.atlassian.net/browse/PM-19603

## 📔 Objective

The goal of this PR is to provide a consistent, clear interface for
asymmetric encryption. For asymmetric encryption we should never encrypt
data, only share keys - "encapsulate". If data access is to be shared to
another account, then this should be done via creating a new symmetric
key, and sharing access to it.

Further, currently our asymmetric data sharing makes no proofs about
authenticity of the sender, that is, the data is unsigned. This makes
that clear on the publicly exposed interface, and leaves the room open
to add an interface that explicitly does provide sender authenticity via
signing.

This PR renames the interfaces to reflect that. Further, we drop the
AsymmetricKey impl of the KeyEncryptable trait, specifically since it
should *not* be able to encrypt arbitrary data, but also because a new
interface that does provide sender authenticity should be the primarily
used interface, which does not line up with the encryptable trait.

## ⏰ Reminders before review

- Contributor guidelines followed
- All formatters and local linters executed and passed
- Written new unit and / or integration tests where applicable
- Protected functional changes with optionality (feature flags)
- Used internationalization (i18n) for all UI strings
- CI builds passed
- Communicated to DevOps any deployment requirements
- Updated any necessary documentation (Confluence, contributing docs) or
informed the documentation
  team

## 🦮 Reviewer guidelines

<!-- Suggested interactions but feel free to use (or not) as you desire!
-->

- 👍 (`:+1:`) or similar for great changes
- 📝 (`:memo:`) or ℹ️ (`:information_source:`) for notes or general info
- ❓ (`:question:`) for questions
- 🤔 (`:thinking:`) or 💭 (`:thought_balloon:`) for more open inquiry
that's not quite a confirmed
  issue and could potentially benefit from discussion
- 🎨 (`:art:`) for suggestions / improvements
- ❌ (`:x:`) or ⚠️ (`:warning:`) for more significant problems or
concerns needing attention
- 🌱 (`:seedling:`) or ♻️ (`:recycle:`) for future improvements or
indications of technical debt
- ⛏ (`:pick:`) for minor or nitpick changes