Fix SM init crypto (#260)
## 🎟️ Tracking

<!-- Paste the link to the Jira or GitHub issue or otherwise describe /
point to where this change is coming from. -->

## 📔 Objective

Before the crypto refactor a user key was required to initialize crypto.
Because secrets manager only has a single organization key, we used to
set it as the user key and internally whenever an organization key was
requested we'd return the user key if there was no private key.

With the crypto refactor, we removed this user key requirement and the
code fallback when selecting a key, but we didn't change the
initialization method to save the key as an organization key.

This PR changes the secrets manager initialization to do the correct
thing here, and also brings a mocked test that was previously disabled
in core to the bitwarden-sm crate. I've also fixed some compile time
features that were incorrect and were causing compilation issues when
running tests on bitwarden-sm.

## ⏰ Reminders before review

- Contributor guidelines followed
- All formatters and local linters executed and passed
- Written new unit and / or integration tests where applicable
- Protected functional changes with optionality (feature flags)
- Used internationalization (i18n) for all UI strings
- CI builds passed
- Communicated to DevOps any deployment requirements
- Updated any necessary documentation (Confluence, contributing docs) or
informed the documentation
  team

## 🦮 Reviewer guidelines

<!-- Suggested interactions but feel free to use (or not) as you desire!
-->

- 👍 (`:+1:`) or similar for great changes
- 📝 (`:memo:`) or ℹ️ (`:information_source:`) for notes or general info
- ❓ (`:question:`) for questions
- 🤔 (`:thinking:`) or 💭 (`:thought_balloon:`) for more open inquiry
that's not quite a confirmed
  issue and could potentially benefit from discussion
- 🎨 (`:art:`) for suggestions / improvements
- ❌ (`:x:`) or ⚠️ (`:warning:`) for more significant problems or
concerns needing attention
- 🌱 (`:seedling:`) or ♻️ (`:recycle:`) for future improvements or
indications of technical debt
- ⛏ (`:pick:`) for minor or nitpick changes