[PM-20957] Update UniFFI to 0.29.1 (#256)
## 🎟️ Tracking

https://bitwarden.atlassian.net/browse/PM-20957

## 📔 Objective

This replaces the renovate PR as the changes are fairly big.

Update UniFFI to version 0.29.1. This requires bumping MSRV to 1.82, as
that's the requirement from UniFFI.

The migration consists of a few parts:
- Replaced all the uses of `UniffiCustomTypeConverter` by the simpler
`custom_type!` macro. Note that for types defined in other crates, we
still use the same macro, but with the `remote,` value added.
- Reworked all the uses of `ffi_converter_forward!`:
- If they were forwarding bitwarden defined types, they aren't needed
anymore.
- If they were forwarding types from third party crates, they were
replaced by `use_remote_type!`
- Replaced a usage of `uniffi::deps::log` by the `log` crate directly.
- While replacing the implementation of `LinkedId`, I've noticed that we
had two implementations of `TryFrom<U32>`, the one used by uniffi based
on deserialize, and a hardcoded map. I've combined them in one using the
deserialize strategy.

I've tested this on the Android and iOS native apps, and they both
compile and decrypt the vault without issues.

## ⏰ Reminders before review

- Contributor guidelines followed
- All formatters and local linters executed and passed
- Written new unit and / or integration tests where applicable
- Protected functional changes with optionality (feature flags)
- Used internationalization (i18n) for all UI strings
- CI builds passed
- Communicated to DevOps any deployment requirements
- Updated any necessary documentation (Confluence, contributing docs) or
informed the documentation
  team

## 🦮 Reviewer guidelines

<!-- Suggested interactions but feel free to use (or not) as you desire!
-->

- 👍 (`:+1:`) or similar for great changes
- 📝 (`:memo:`) or ℹ️ (`:information_source:`) for notes or general info
- ❓ (`:question:`) for questions
- 🤔 (`:thinking:`) or 💭 (`:thought_balloon:`) for more open inquiry
that's not quite a confirmed
  issue and could potentially benefit from discussion
- 🎨 (`:art:`) for suggestions / improvements
- ❌ (`:x:`) or ⚠️ (`:warning:`) for more significant problems or
concerns needing attention
- 🌱 (`:seedling:`) or ♻️ (`:recycle:`) for future improvements or
indications of technical debt
- ⛏ (`:pick:`) for minor or nitpick changes