Warn on missing doc (#199)
We want to encourage documentation on all public items. I'm not
completely sold on the idea since certain things are self-descriptive
but we also should strongly discourage undocumented things.

Adds `missing_docs = "warn"` to cargo workspace and adds
`allow(missing_docs)` to all existing violations. The plan is to prevent
new infractions and to slowly document as we work on areas.