[PM-22647] Move build subtitle into cipher types (#307)
`Cipher` is currently non idiomatic rust due to a desire to simplify the
conversion layers and to accurately replicate existing behaviour. This
has some disadvantages such as `Cipher` having to own all business logic
even if it really belongs to one of the cipher types.

This PR introduces a `CipherKind` enum which should be implemented by
all cipher types. It also introduces a `get_kind` helper on `Cipher`
which resolves the accurate `CipherKind` depending on cipher type. This
allows us to delegate implementations to the relevant cipher kind
avoiding complex match statements in `Cipher`.