`CipherListView` updates (#306)
## 🎟️ Tracking

Tangential to
[PM-22134](https://bitwarden.atlassian.net/browse/PM-22134)

## 📔 Objective

Two updates to ensure parity when migrating from `CipherView` to
`CipherListView` on the clients repo
- Added `local_data` so the type propagates to the clients repo and can
be set there.
- Added  `CopiableCipherFields` enum + property.
- In the browser the copy menu / actions are dynamic based on the
populated fields on the cipher. The `CipherListView` doesn't contain
these details. This array is filled with the available fields so the
extension can still check their availability. When a value is needed the
full cipher is decrypted.
- Added `has_old_attachments` to identify attachments following old
encryption methods. This populates UI signaling to the user that the
attachment needs to be re-encrypted.

## 🦮 Reviewer guidelines

<!-- Suggested interactions but feel free to use (or not) as you desire!
-->

- 👍 (`:+1:`) or similar for great changes
- 📝 (`:memo:`) or ℹ️ (`:information_source:`) for notes or general info
- ❓ (`:question:`) for questions
- 🤔 (`:thinking:`) or 💭 (`:thought_balloon:`) for more open inquiry
that's not quite a confirmed
  issue and could potentially benefit from discussion
- 🎨 (`:art:`) for suggestions / improvements
- ❌ (`:x:`) or ⚠️ (`:warning:`) for more significant problems or
concerns needing attention
- 🌱 (`:seedling:`) or ♻️ (`:recycle:`) for future improvements or
indications of technical debt
- ⛏ (`:pick:`) for minor or nitpick changes


[PM-22134]:
https://bitwarden.atlassian.net/browse/PM-22134?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ