[PM-20614] Cose Content Type (#203)
## 🎟️ Tracking

https://bitwarden.atlassian.net/browse/PM-20614

## 📔 Objective

Encrypt type 7 (CoseEncrypt0 with currently Xchacha20poly1305) requires
specifying a content format. Since the encrypted content is just a byte
array, this alone is not enough to understand what format the data
contained has. In some cases, multiple formats may be OK.

These cases include:
- Private keys
  - PKCS8 DER / CoseKey
- Symmetric keys
- Bitwarden Custom Key Byte Serialization ("just key bytes concatenated
(and padding if it is a CoseKey)"), CoseKey
- Utf8 for most cipher fields
- Future:
  - Blob Cipher: Json/Cbor?

To distinguish between these, we need to introduce a content format that
is passed through until the encryption happens. For AES256-CBC-HMAC, it
gets ignored, for CoseEncrypt0/Xchacha20poly1305, it gets written as a
protected header onto the encstring/encrypt0 object. We do not add
validation on the decryption path in this PR.

For key wrapping, when wrapping cosekeys with cosekeys, the Bitwarden
Custom Key Byte Serialization with padding is not used, but the raw
CoseKey without padding is used; but can be distinguished from other
keys on unwrapping by inspecting the content format.

Further to ensure the content format is passed along, the encryptable
trait is split into 3 traits by this PR.
`CompositeEncryptable` is a trait that represents encryption of
composite objects. Composite objects here are objects that have
sub-encyrptables under the same key. For instance, a cipher is a
composite encryptable, since all fields are individually encrypted. It
also has some sub-composite encryptables (Uri). Composite encryptables
only defer encryption to one of the other traits, and do not themselves
encrypt directly, and thus also don't have a content type.

For the Utf8 content type specifically, for XChaCha20Poly1305
encryption, a padding scheme is added, in order to hide the direct
relationship of ciphertext length to plaintext length. The content type
written onto the encrypt0 object is a custom type to indicate this.

The since bytes `Vec<u8>`. alone are not enough to carry a content type,
a new generic struct `Bytes<ConstContentType>` is introduced and
returned in various places such as:
- DER private key
- DER public key
- Symmetric crypto key

The `ConstContentType` is sealed and not implementable by other crates,
and type aliases are provided for more convenient / less cognitive
overhead syntax.

For these, the EncryptableWithContentType trait(s) are implemented.

## ⏰ Reminders before review

- Contributor guidelines followed
- All formatters and local linters executed and passed
- Written new unit and / or integration tests where applicable
- Protected functional changes with optionality (feature flags)
- Used internationalization (i18n) for all UI strings
- CI builds passed
- Communicated to DevOps any deployment requirements
- Updated any necessary documentation (Confluence, contributing docs) or
informed the documentation
  team

## 🦮 Reviewer guidelines

<!-- Suggested interactions but feel free to use (or not) as you desire!
-->

- 👍 (`:+1:`) or similar for great changes
- 📝 (`:memo:`) or ℹ️ (`:information_source:`) for notes or general info
- ❓ (`:question:`) for questions
- 🤔 (`:thinking:`) or 💭 (`:thought_balloon:`) for more open inquiry
that's not quite a confirmed
  issue and could potentially benefit from discussion
- 🎨 (`:art:`) for suggestions / improvements
- ❌ (`:x:`) or ⚠️ (`:warning:`) for more significant problems or
concerns needing attention
- 🌱 (`:seedling:`) or ♻️ (`:recycle:`) for future improvements or
indications of technical debt
- ⛏ (`:pick:`) for minor or nitpick changes

---------

Co-authored-by: Matt Gibson <mgibson@bitwarden.com>
Co-authored-by: Oscar Hinton <Hinton@users.noreply.github.com>